@@ -1,29 +1,30 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
-// source: fabric.proto
+// versions:
+// 	protoc-gen-go v1.25.0
+// 	protoc        v3.11.4
+// source: fabric/fabric.proto
 
 package fabric
 
 import (
-	context "context"
-	fmt "fmt"
 	proto "github.com/golang/protobuf/proto"
 	_ "google.golang.org/genproto/googleapis/api/annotations"
-	grpc "google.golang.org/grpc"
-	codes "google.golang.org/grpc/codes"
-	status "google.golang.org/grpc/status"
-	math "math"
+	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
+	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
+	reflect "reflect"
+	sync "sync"
 )
 
-// Reference imports to suppress errors if they are not otherwise used.
-var _ = proto.Marshal
-var _ = fmt.Errorf
-var _ = math.Inf
+const (
+	// Verify that this generated code is sufficiently up-to-date.
+	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
+	// Verify that runtime/protoimpl is sufficiently up-to-date.
+	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
+)
 
-// This is a compile-time assertion to ensure that this generated file
-// is compatible with the proto package it is being compiled against.
-// A compilation error at this line likely means your copy of the
-// proto package needs to be updated.
-const _ = proto.ProtoPackageIsVersion3 // please upgrade the proto package
+// This is a compile-time assertion that a sufficiently up-to-date version
+// of the legacy proto package is being used.
+const _ = proto.ProtoPackageIsVersion4
 
 // Type of Fabric transaction: INVOKE or QUERY.
 type TransactionType int32
@@ -33,70 +34,103 @@ const (
 	TransactionType_QUERY  TransactionType = 1
 )
 
-var TransactionType_name = map[int32]string{
-	0: "INVOKE",
-	1: "QUERY",
-}
+// Enum value maps for TransactionType.
+var (
+	TransactionType_name = map[int32]string{
+		0: "INVOKE",
+		1: "QUERY",
+	}
+	TransactionType_value = map[string]int32{
+		"INVOKE": 0,
+		"QUERY":  1,
+	}
+)
 
-var TransactionType_value = map[string]int32{
-	"INVOKE": 0,
-	"QUERY":  1,
+func (x TransactionType) Enum() *TransactionType {
+	p := new(TransactionType)
+	*p = x
+	return p
 }
 
 func (x TransactionType) String() string {
-	return proto.EnumName(TransactionType_name, int32(x))
+	return protoimpl.X.EnumStringOf(x.Descriptor(), protoreflect.EnumNumber(x))
 }
 
+func (TransactionType) Descriptor() protoreflect.EnumDescriptor {
+	return file_fabric_fabric_proto_enumTypes[0].Descriptor()
+}
+
+func (TransactionType) Type() protoreflect.EnumType {
+	return &file_fabric_fabric_proto_enumTypes[0]
+}
+
+func (x TransactionType) Number() protoreflect.EnumNumber {
+	return protoreflect.EnumNumber(x)
+}
+
+// Deprecated: Use TransactionType.Descriptor instead.
 func (TransactionType) EnumDescriptor() ([]byte, []int) {
-	return fileDescriptor_284efff686d8e9bf, []int{0}
+	return file_fabric_fabric_proto_rawDescGZIP(), []int{0}
 }
 
 // Fabric transaction request.
 type TransactionRequest struct {
-	Data                 *TransactionData `protobuf:"bytes,1,opt,name=data,proto3" json:"data,omitempty"`
-	XXX_NoUnkeyedLiteral struct{}         `json:"-"`
-	XXX_unrecognized     []byte           `json:"-"`
-	XXX_sizecache        int32            `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Data *TransactionData `protobuf:"bytes,1,opt,name=data,proto3" json:"data,omitempty"`
 }
 
-func (m *TransactionRequest) Reset()         { *m = TransactionRequest{} }
-func (m *TransactionRequest) String() string { return proto.CompactTextString(m) }
-func (*TransactionRequest) ProtoMessage()    {}
+func (x *TransactionRequest) Reset() {
+	*x = TransactionRequest{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_fabric_fabric_proto_msgTypes[0]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *TransactionRequest) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*TransactionRequest) ProtoMessage() {}
+
+func (x *TransactionRequest) ProtoReflect() protoreflect.Message {
+	mi := &file_fabric_fabric_proto_msgTypes[0]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use TransactionRequest.ProtoReflect.Descriptor instead.
 func (*TransactionRequest) Descriptor() ([]byte, []int) {
-	return fileDescriptor_284efff686d8e9bf, []int{0}
+	return file_fabric_fabric_proto_rawDescGZIP(), []int{0}
 }
 
-func (m *TransactionRequest) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_TransactionRequest.Unmarshal(m, b)
-}
-func (m *TransactionRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_TransactionRequest.Marshal(b, m, deterministic)
-}
-func (m *TransactionRequest) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_TransactionRequest.Merge(m, src)
-}
-func (m *TransactionRequest) XXX_Size() int {
-	return xxx_messageInfo_TransactionRequest.Size(m)
-}
-func (m *TransactionRequest) XXX_DiscardUnknown() {
-	xxx_messageInfo_TransactionRequest.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_TransactionRequest proto.InternalMessageInfo
-
-func (m *TransactionRequest) GetData() *TransactionData {
-	if m != nil {
-		return m.Data
+func (x *TransactionRequest) GetData() *TransactionData {
+	if x != nil {
+		return x.Data
 	}
 	return nil
 }
 
 // Request data for a Fabric chaincode transaction
 type TransactionData struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
 	// Required. ID of a Fabric connection returned by a ConnectRequest
 	ConnectionId uint64 `protobuf:"varint,1,opt,name=connection_id,json=connectionId,proto3" json:"connection_id,omitempty"`
 	// Required. 0=INVOKE or 1=QUERY
-	Type TransactionType `protobuf:"varint,2,opt,name=type,proto3,enum=dovetail.fabric.v1.TransactionType" json:"type,omitempty"`
+	Type TransactionType `protobuf:"varint,2,opt,name=type,proto3,enum=fabric.TransactionType" json:"type,omitempty"`
 	// Required. Name of the Fabric chaincode to be called.
 	ChaincodeId string `protobuf:"bytes,3,opt,name=chaincode_id,json=chaincodeId,proto3" json:"chaincode_id,omitempty"`
 	// Request timeout in milliseconds. Default is 0, i.e., no timeout.
@@ -108,194 +142,218 @@ type TransactionData struct {
 	// parameters for the transaction
 	Parameter []string `protobuf:"bytes,7,rep,name=parameter,proto3" json:"parameter,omitempty"`
 	// transient map for the transaction as a JSON string.
-	TransientMap         string   `protobuf:"bytes,8,opt,name=transient_map,json=transientMap,proto3" json:"transient_map,omitempty"`
-	XXX_NoUnkeyedLiteral struct{} `json:"-"`
-	XXX_unrecognized     []byte   `json:"-"`
-	XXX_sizecache        int32    `json:"-"`
+	TransientMap string `protobuf:"bytes,8,opt,name=transient_map,json=transientMap,proto3" json:"transient_map,omitempty"`
 }
 
-func (m *TransactionData) Reset()         { *m = TransactionData{} }
-func (m *TransactionData) String() string { return proto.CompactTextString(m) }
-func (*TransactionData) ProtoMessage()    {}
+func (x *TransactionData) Reset() {
+	*x = TransactionData{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_fabric_fabric_proto_msgTypes[1]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *TransactionData) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*TransactionData) ProtoMessage() {}
+
+func (x *TransactionData) ProtoReflect() protoreflect.Message {
+	mi := &file_fabric_fabric_proto_msgTypes[1]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use TransactionData.ProtoReflect.Descriptor instead.
 func (*TransactionData) Descriptor() ([]byte, []int) {
-	return fileDescriptor_284efff686d8e9bf, []int{1}
+	return file_fabric_fabric_proto_rawDescGZIP(), []int{1}
 }
 
-func (m *TransactionData) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_TransactionData.Unmarshal(m, b)
-}
-func (m *TransactionData) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_TransactionData.Marshal(b, m, deterministic)
-}
-func (m *TransactionData) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_TransactionData.Merge(m, src)
-}
-func (m *TransactionData) XXX_Size() int {
-	return xxx_messageInfo_TransactionData.Size(m)
-}
-func (m *TransactionData) XXX_DiscardUnknown() {
-	xxx_messageInfo_TransactionData.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_TransactionData proto.InternalMessageInfo
-
-func (m *TransactionData) GetConnectionId() uint64 {
-	if m != nil {
-		return m.ConnectionId
+func (x *TransactionData) GetConnectionId() uint64 {
+	if x != nil {
+		return x.ConnectionId
 	}
 	return 0
 }
 
-func (m *TransactionData) GetType() TransactionType {
-	if m != nil {
-		return m.Type
+func (x *TransactionData) GetType() TransactionType {
+	if x != nil {
+		return x.Type
 	}
 	return TransactionType_INVOKE
 }
 
-func (m *TransactionData) GetChaincodeId() string {
-	if m != nil {
-		return m.ChaincodeId
+func (x *TransactionData) GetChaincodeId() string {
+	if x != nil {
+		return x.ChaincodeId
 	}
 	return ""
 }
 
-func (m *TransactionData) GetTimeout() int64 {
-	if m != nil {
-		return m.Timeout
+func (x *TransactionData) GetTimeout() int64 {
+	if x != nil {
+		return x.Timeout
 	}
 	return 0
 }
 
-func (m *TransactionData) GetEndpoint() []string {
-	if m != nil {
-		return m.Endpoint
+func (x *TransactionData) GetEndpoint() []string {
+	if x != nil {
+		return x.Endpoint
 	}
 	return nil
 }
 
-func (m *TransactionData) GetTransaction() string {
-	if m != nil {
-		return m.Transaction
+func (x *TransactionData) GetTransaction() string {
+	if x != nil {
+		return x.Transaction
 	}
 	return ""
 }
 
-func (m *TransactionData) GetParameter() []string {
-	if m != nil {
-		return m.Parameter
+func (x *TransactionData) GetParameter() []string {
+	if x != nil {
+		return x.Parameter
 	}
 	return nil
 }
 
-func (m *TransactionData) GetTransientMap() string {
-	if m != nil {
-		return m.TransientMap
+func (x *TransactionData) GetTransientMap() string {
+	if x != nil {
+		return x.TransientMap
 	}
 	return ""
 }
 
 // Response from a Fabric transaction
 type TransactionResponse struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
 	// Status code. 200=success
 	Code int32 `protobuf:"varint,1,opt,name=code,proto3" json:"code,omitempty"`
 	// Error messages if code is not 200
 	Message string `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"`
 	// Returned data as JSON string
-	Data                 string   `protobuf:"bytes,3,opt,name=data,proto3" json:"data,omitempty"`
-	XXX_NoUnkeyedLiteral struct{} `json:"-"`
-	XXX_unrecognized     []byte   `json:"-"`
-	XXX_sizecache        int32    `json:"-"`
+	Data string `protobuf:"bytes,3,opt,name=data,proto3" json:"data,omitempty"`
 }
 
-func (m *TransactionResponse) Reset()         { *m = TransactionResponse{} }
-func (m *TransactionResponse) String() string { return proto.CompactTextString(m) }
-func (*TransactionResponse) ProtoMessage()    {}
+func (x *TransactionResponse) Reset() {
+	*x = TransactionResponse{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_fabric_fabric_proto_msgTypes[2]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *TransactionResponse) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*TransactionResponse) ProtoMessage() {}
+
+func (x *TransactionResponse) ProtoReflect() protoreflect.Message {
+	mi := &file_fabric_fabric_proto_msgTypes[2]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use TransactionResponse.ProtoReflect.Descriptor instead.
 func (*TransactionResponse) Descriptor() ([]byte, []int) {
-	return fileDescriptor_284efff686d8e9bf, []int{2}
+	return file_fabric_fabric_proto_rawDescGZIP(), []int{2}
 }
 
-func (m *TransactionResponse) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_TransactionResponse.Unmarshal(m, b)
-}
-func (m *TransactionResponse) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_TransactionResponse.Marshal(b, m, deterministic)
-}
-func (m *TransactionResponse) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_TransactionResponse.Merge(m, src)
-}
-func (m *TransactionResponse) XXX_Size() int {
-	return xxx_messageInfo_TransactionResponse.Size(m)
-}
-func (m *TransactionResponse) XXX_DiscardUnknown() {
-	xxx_messageInfo_TransactionResponse.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_TransactionResponse proto.InternalMessageInfo
-
-func (m *TransactionResponse) GetCode() int32 {
-	if m != nil {
-		return m.Code
+func (x *TransactionResponse) GetCode() int32 {
+	if x != nil {
+		return x.Code
 	}
 	return 0
 }
 
-func (m *TransactionResponse) GetMessage() string {
-	if m != nil {
-		return m.Message
+func (x *TransactionResponse) GetMessage() string {
+	if x != nil {
+		return x.Message
 	}
 	return ""
 }
 
-func (m *TransactionResponse) GetData() string {
-	if m != nil {
-		return m.Data
+func (x *TransactionResponse) GetData() string {
+	if x != nil {
+		return x.Data
 	}
 	return ""
 }
 
 // Fabric connection request.
 type ConnectionRequest struct {
-	Data                 *ConnectionData `protobuf:"bytes,1,opt,name=data,proto3" json:"data,omitempty"`
-	XXX_NoUnkeyedLiteral struct{}        `json:"-"`
-	XXX_unrecognized     []byte          `json:"-"`
-	XXX_sizecache        int32           `json:"-"`
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Data *ConnectionData `protobuf:"bytes,1,opt,name=data,proto3" json:"data,omitempty"`
 }
 
-func (m *ConnectionRequest) Reset()         { *m = ConnectionRequest{} }
-func (m *ConnectionRequest) String() string { return proto.CompactTextString(m) }
-func (*ConnectionRequest) ProtoMessage()    {}
+func (x *ConnectionRequest) Reset() {
+	*x = ConnectionRequest{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_fabric_fabric_proto_msgTypes[3]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *ConnectionRequest) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*ConnectionRequest) ProtoMessage() {}
+
+func (x *ConnectionRequest) ProtoReflect() protoreflect.Message {
+	mi := &file_fabric_fabric_proto_msgTypes[3]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use ConnectionRequest.ProtoReflect.Descriptor instead.
 func (*ConnectionRequest) Descriptor() ([]byte, []int) {
-	return fileDescriptor_284efff686d8e9bf, []int{3}
+	return file_fabric_fabric_proto_rawDescGZIP(), []int{3}
 }
 
-func (m *ConnectionRequest) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_ConnectionRequest.Unmarshal(m, b)
-}
-func (m *ConnectionRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_ConnectionRequest.Marshal(b, m, deterministic)
-}
-func (m *ConnectionRequest) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_ConnectionRequest.Merge(m, src)
-}
-func (m *ConnectionRequest) XXX_Size() int {
-	return xxx_messageInfo_ConnectionRequest.Size(m)
-}
-func (m *ConnectionRequest) XXX_DiscardUnknown() {
-	xxx_messageInfo_ConnectionRequest.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_ConnectionRequest proto.InternalMessageInfo
-
-func (m *ConnectionRequest) GetData() *ConnectionData {
-	if m != nil {
-		return m.Data
+func (x *ConnectionRequest) GetData() *ConnectionData {
+	if x != nil {
+		return x.Data
 	}
 	return nil
 }
 
 // Request data for creating a Fabric connection in the gateway service
 type ConnectionData struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
 	// Name of the Fabric channel to connect to. Default is pre-configured by gateway service.
 	ChannelId string `protobuf:"bytes,1,opt,name=channel_id,json=channelId,proto3" json:"channel_id,omitempty"`
 	// Name of Fabric client user for establishing the connection. Default is pre-configured by gateway service.
@@ -305,304 +363,354 @@ type ConnectionData struct {
 	// File name for Fabric network config. Default is pre-configured by gateway service.
 	NetworkConfig string `protobuf:"bytes,4,opt,name=network_config,json=networkConfig,proto3" json:"network_config,omitempty"`
 	// File name for endpoint patten matchers. Default is pre-configured by gateway service.
-	PattenMatchers       string   `protobuf:"bytes,5,opt,name=patten_matchers,json=pattenMatchers,proto3" json:"patten_matchers,omitempty"`
-	XXX_NoUnkeyedLiteral struct{} `json:"-"`
-	XXX_unrecognized     []byte   `json:"-"`
-	XXX_sizecache        int32    `json:"-"`
+	PattenMatchers string `protobuf:"bytes,5,opt,name=patten_matchers,json=pattenMatchers,proto3" json:"patten_matchers,omitempty"`
 }
 
-func (m *ConnectionData) Reset()         { *m = ConnectionData{} }
-func (m *ConnectionData) String() string { return proto.CompactTextString(m) }
-func (*ConnectionData) ProtoMessage()    {}
+func (x *ConnectionData) Reset() {
+	*x = ConnectionData{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_fabric_fabric_proto_msgTypes[4]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *ConnectionData) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*ConnectionData) ProtoMessage() {}
+
+func (x *ConnectionData) ProtoReflect() protoreflect.Message {
+	mi := &file_fabric_fabric_proto_msgTypes[4]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use ConnectionData.ProtoReflect.Descriptor instead.
 func (*ConnectionData) Descriptor() ([]byte, []int) {
-	return fileDescriptor_284efff686d8e9bf, []int{4}
+	return file_fabric_fabric_proto_rawDescGZIP(), []int{4}
 }
 
-func (m *ConnectionData) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_ConnectionData.Unmarshal(m, b)
-}
-func (m *ConnectionData) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_ConnectionData.Marshal(b, m, deterministic)
-}
-func (m *ConnectionData) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_ConnectionData.Merge(m, src)
-}
-func (m *ConnectionData) XXX_Size() int {
-	return xxx_messageInfo_ConnectionData.Size(m)
-}
-func (m *ConnectionData) XXX_DiscardUnknown() {
-	xxx_messageInfo_ConnectionData.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_ConnectionData proto.InternalMessageInfo
-
-func (m *ConnectionData) GetChannelId() string {
-	if m != nil {
-		return m.ChannelId
+func (x *ConnectionData) GetChannelId() string {
+	if x != nil {
+		return x.ChannelId
 	}
 	return ""
 }
 
-func (m *ConnectionData) GetUserName() string {
-	if m != nil {
-		return m.UserName
+func (x *ConnectionData) GetUserName() string {
+	if x != nil {
+		return x.UserName
 	}
 	return ""
 }
 
-func (m *ConnectionData) GetOrgName() string {
-	if m != nil {
-		return m.OrgName
+func (x *ConnectionData) GetOrgName() string {
+	if x != nil {
+		return x.OrgName
 	}
 	return ""
 }
 
-func (m *ConnectionData) GetNetworkConfig() string {
-	if m != nil {
-		return m.NetworkConfig
+func (x *ConnectionData) GetNetworkConfig() string {
+	if x != nil {
+		return x.NetworkConfig
 	}
 	return ""
 }
 
-func (m *ConnectionData) GetPattenMatchers() string {
-	if m != nil {
-		return m.PattenMatchers
+func (x *ConnectionData) GetPattenMatchers() string {
+	if x != nil {
+		return x.PattenMatchers
 	}
 	return ""
 }
 
 // Response from a Fabric connection request
 type ConnectionResponse struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
 	// Status code. 200=success
 	Code int32 `protobuf:"varint,1,opt,name=code,proto3" json:"code,omitempty"`
 	// Error messages if code is not 200
 	Message string `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"`
 	// ID of the resulting Fabric connection
-	ConnectionId         uint64   `protobuf:"varint,3,opt,name=connection_id,json=connectionId,proto3" json:"connection_id,omitempty"`
-	XXX_NoUnkeyedLiteral struct{} `json:"-"`
-	XXX_unrecognized     []byte   `json:"-"`
-	XXX_sizecache        int32    `json:"-"`
+	ConnectionId uint64 `protobuf:"varint,3,opt,name=connection_id,json=connectionId,proto3" json:"connection_id,omitempty"`
 }
 
-func (m *ConnectionResponse) Reset()         { *m = ConnectionResponse{} }
-func (m *ConnectionResponse) String() string { return proto.CompactTextString(m) }
-func (*ConnectionResponse) ProtoMessage()    {}
+func (x *ConnectionResponse) Reset() {
+	*x = ConnectionResponse{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_fabric_fabric_proto_msgTypes[5]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *ConnectionResponse) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*ConnectionResponse) ProtoMessage() {}
+
+func (x *ConnectionResponse) ProtoReflect() protoreflect.Message {
+	mi := &file_fabric_fabric_proto_msgTypes[5]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use ConnectionResponse.ProtoReflect.Descriptor instead.
 func (*ConnectionResponse) Descriptor() ([]byte, []int) {
-	return fileDescriptor_284efff686d8e9bf, []int{5}
+	return file_fabric_fabric_proto_rawDescGZIP(), []int{5}
 }
 
-func (m *ConnectionResponse) XXX_Unmarshal(b []byte) error {
-	return xxx_messageInfo_ConnectionResponse.Unmarshal(m, b)
-}
-func (m *ConnectionResponse) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
-	return xxx_messageInfo_ConnectionResponse.Marshal(b, m, deterministic)
-}
-func (m *ConnectionResponse) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_ConnectionResponse.Merge(m, src)
-}
-func (m *ConnectionResponse) XXX_Size() int {
-	return xxx_messageInfo_ConnectionResponse.Size(m)
-}
-func (m *ConnectionResponse) XXX_DiscardUnknown() {
-	xxx_messageInfo_ConnectionResponse.DiscardUnknown(m)
-}
-
-var xxx_messageInfo_ConnectionResponse proto.InternalMessageInfo
-
-func (m *ConnectionResponse) GetCode() int32 {
-	if m != nil {
-		return m.Code
+func (x *ConnectionResponse) GetCode() int32 {
+	if x != nil {
+		return x.Code
 	}
 	return 0
 }
 
-func (m *ConnectionResponse) GetMessage() string {
-	if m != nil {
-		return m.Message
+func (x *ConnectionResponse) GetMessage() string {
+	if x != nil {
+		return x.Message
 	}
 	return ""
 }
 
-func (m *ConnectionResponse) GetConnectionId() uint64 {
-	if m != nil {
-		return m.ConnectionId
+func (x *ConnectionResponse) GetConnectionId() uint64 {
+	if x != nil {
+		return x.ConnectionId
 	}
 	return 0
 }
 
-func init() {
-	proto.RegisterEnum("dovetail.fabric.v1.TransactionType", TransactionType_name, TransactionType_value)
-	proto.RegisterType((*TransactionRequest)(nil), "dovetail.fabric.v1.TransactionRequest")
-	proto.RegisterType((*TransactionData)(nil), "dovetail.fabric.v1.TransactionData")
-	proto.RegisterType((*TransactionResponse)(nil), "dovetail.fabric.v1.TransactionResponse")
-	proto.RegisterType((*ConnectionRequest)(nil), "dovetail.fabric.v1.ConnectionRequest")
-	proto.RegisterType((*ConnectionData)(nil), "dovetail.fabric.v1.ConnectionData")
-	proto.RegisterType((*ConnectionResponse)(nil), "dovetail.fabric.v1.ConnectionResponse")
+var File_fabric_fabric_proto protoreflect.FileDescriptor
+
+var file_fabric_fabric_proto_rawDesc = []byte{
+	0x0a, 0x13, 0x66, 0x61, 0x62, 0x72, 0x69, 0x63, 0x2f, 0x66, 0x61, 0x62, 0x72, 0x69, 0x63, 0x2e,
+	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x06, 0x66, 0x61, 0x62, 0x72, 0x69, 0x63, 0x1a, 0x1c, 0x67,
+	0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x61, 0x70, 0x69, 0x2f, 0x61, 0x6e, 0x6e, 0x6f, 0x74, 0x61,
+	0x74, 0x69, 0x6f, 0x6e, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x41, 0x0a, 0x12, 0x54,
+	0x72, 0x61, 0x6e, 0x73, 0x61, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
+	0x74, 0x12, 0x2b, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x61, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32,
+	0x17, 0x2e, 0x66, 0x61, 0x62, 0x72, 0x69, 0x63, 0x2e, 0x54, 0x72, 0x61, 0x6e, 0x73, 0x61, 0x63,
+	0x74, 0x69, 0x6f, 0x6e, 0x44, 0x61, 0x74, 0x61, 0x52, 0x04, 0x64, 0x61, 0x74, 0x61, 0x22, 0xa1,
+	0x02, 0x0a, 0x0f, 0x54, 0x72, 0x61, 0x6e, 0x73, 0x61, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x44, 0x61,
+	0x74, 0x61, 0x12, 0x23, 0x0a, 0x0d, 0x63, 0x6f, 0x6e, 0x6e, 0x65, 0x63, 0x74, 0x69, 0x6f, 0x6e,
+	0x5f, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x04, 0x52, 0x0c, 0x63, 0x6f, 0x6e, 0x6e, 0x65,
+	0x63, 0x74, 0x69, 0x6f, 0x6e, 0x49, 0x64, 0x12, 0x2b, 0x0a, 0x04, 0x74, 0x79, 0x70, 0x65, 0x18,
+	0x02, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x17, 0x2e, 0x66, 0x61, 0x62, 0x72, 0x69, 0x63, 0x2e, 0x54,
+	0x72, 0x61, 0x6e, 0x73, 0x61, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x54, 0x79, 0x70, 0x65, 0x52, 0x04,
+	0x74, 0x79, 0x70, 0x65, 0x12, 0x21, 0x0a, 0x0c, 0x63, 0x68, 0x61, 0x69, 0x6e, 0x63, 0x6f, 0x64,
+	0x65, 0x5f, 0x69, 0x64, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0b, 0x63, 0x68, 0x61, 0x69,
+	0x6e, 0x63, 0x6f, 0x64, 0x65, 0x49, 0x64, 0x12, 0x18, 0x0a, 0x07, 0x74, 0x69, 0x6d, 0x65, 0x6f,
+	0x75, 0x74, 0x18, 0x04, 0x20, 0x01, 0x28, 0x03, 0x52, 0x07, 0x74, 0x69, 0x6d, 0x65, 0x6f, 0x75,
+	0x74, 0x12, 0x1a, 0x0a, 0x08, 0x65, 0x6e, 0x64, 0x70, 0x6f, 0x69, 0x6e, 0x74, 0x18, 0x05, 0x20,
+	0x03, 0x28, 0x09, 0x52, 0x08, 0x65, 0x6e, 0x64, 0x70, 0x6f, 0x69, 0x6e, 0x74, 0x12, 0x20, 0x0a,
+	0x0b, 0x74, 0x72, 0x61, 0x6e, 0x73, 0x61, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x18, 0x06, 0x20, 0x01,
+	0x28, 0x09, 0x52, 0x0b, 0x74, 0x72, 0x61, 0x6e, 0x73, 0x61, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x12,
+	0x1c, 0x0a, 0x09, 0x70, 0x61, 0x72, 0x61, 0x6d, 0x65, 0x74, 0x65, 0x72, 0x18, 0x07, 0x20, 0x03,
+	0x28, 0x09, 0x52, 0x09, 0x70, 0x61, 0x72, 0x61, 0x6d, 0x65, 0x74, 0x65, 0x72, 0x12, 0x23, 0x0a,
+	0x0d, 0x74, 0x72, 0x61, 0x6e, 0x73, 0x69, 0x65, 0x6e, 0x74, 0x5f, 0x6d, 0x61, 0x70, 0x18, 0x08,
+	0x20, 0x01, 0x28, 0x09, 0x52, 0x0c, 0x74, 0x72, 0x61, 0x6e, 0x73, 0x69, 0x65, 0x6e, 0x74, 0x4d,
+	0x61, 0x70, 0x22, 0x57, 0x0a, 0x13, 0x54, 0x72, 0x61, 0x6e, 0x73, 0x61, 0x63, 0x74, 0x69, 0x6f,
+	0x6e, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x63, 0x6f, 0x64,
+	0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x52, 0x04, 0x63, 0x6f, 0x64, 0x65, 0x12, 0x18, 0x0a,
+	0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07,
+	0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x61, 0x18,
+	0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x64, 0x61, 0x74, 0x61, 0x22, 0x3f, 0x0a, 0x11, 0x43,
+	0x6f, 0x6e, 0x6e, 0x65, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
+	0x12, 0x2a, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x61, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x16,
+	0x2e, 0x66, 0x61, 0x62, 0x72, 0x69, 0x63, 0x2e, 0x43, 0x6f, 0x6e, 0x6e, 0x65, 0x63, 0x74, 0x69,
+	0x6f, 0x6e, 0x44, 0x61, 0x74, 0x61, 0x52, 0x04, 0x64, 0x61, 0x74, 0x61, 0x22, 0xb7, 0x01, 0x0a,
+	0x0e, 0x43, 0x6f, 0x6e, 0x6e, 0x65, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x44, 0x61, 0x74, 0x61, 0x12,
+	0x1d, 0x0a, 0x0a, 0x63, 0x68, 0x61, 0x6e, 0x6e, 0x65, 0x6c, 0x5f, 0x69, 0x64, 0x18, 0x01, 0x20,
+	0x01, 0x28, 0x09, 0x52, 0x09, 0x63, 0x68, 0x61, 0x6e, 0x6e, 0x65, 0x6c, 0x49, 0x64, 0x12, 0x1b,
+	0x0a, 0x09, 0x75, 0x73, 0x65, 0x72, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28,
+	0x09, 0x52, 0x08, 0x75, 0x73, 0x65, 0x72, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x19, 0x0a, 0x08, 0x6f,
+	0x72, 0x67, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x6f,
+	0x72, 0x67, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x25, 0x0a, 0x0e, 0x6e, 0x65, 0x74, 0x77, 0x6f, 0x72,
+	0x6b, 0x5f, 0x63, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0d,
+	0x6e, 0x65, 0x74, 0x77, 0x6f, 0x72, 0x6b, 0x43, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x12, 0x27, 0x0a,
+	0x0f, 0x70, 0x61, 0x74, 0x74, 0x65, 0x6e, 0x5f, 0x6d, 0x61, 0x74, 0x63, 0x68, 0x65, 0x72, 0x73,
+	0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0e, 0x70, 0x61, 0x74, 0x74, 0x65, 0x6e, 0x4d, 0x61,
+	0x74, 0x63, 0x68, 0x65, 0x72, 0x73, 0x22, 0x67, 0x0a, 0x12, 0x43, 0x6f, 0x6e, 0x6e, 0x65, 0x63,
+	0x74, 0x69, 0x6f, 0x6e, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x12, 0x0a, 0x04,
+	0x63, 0x6f, 0x64, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x52, 0x04, 0x63, 0x6f, 0x64, 0x65,
+	0x12, 0x18, 0x0a, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28,
+	0x09, 0x52, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x23, 0x0a, 0x0d, 0x63, 0x6f,
+	0x6e, 0x6e, 0x65, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x5f, 0x69, 0x64, 0x18, 0x03, 0x20, 0x01, 0x28,
+	0x04, 0x52, 0x0c, 0x63, 0x6f, 0x6e, 0x6e, 0x65, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x49, 0x64, 0x2a,
+	0x28, 0x0a, 0x0f, 0x54, 0x72, 0x61, 0x6e, 0x73, 0x61, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x54, 0x79,
+	0x70, 0x65, 0x12, 0x0a, 0x0a, 0x06, 0x49, 0x4e, 0x56, 0x4f, 0x4b, 0x45, 0x10, 0x00, 0x12, 0x09,
+	0x0a, 0x05, 0x51, 0x55, 0x45, 0x52, 0x59, 0x10, 0x01, 0x32, 0xcc, 0x01, 0x0a, 0x07, 0x47, 0x61,
+	0x74, 0x65, 0x77, 0x61, 0x79, 0x12, 0x5e, 0x0a, 0x07, 0x43, 0x6f, 0x6e, 0x6e, 0x65, 0x63, 0x74,
+	0x12, 0x19, 0x2e, 0x66, 0x61, 0x62, 0x72, 0x69, 0x63, 0x2e, 0x43, 0x6f, 0x6e, 0x6e, 0x65, 0x63,
+	0x74, 0x69, 0x6f, 0x6e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1a, 0x2e, 0x66, 0x61,
+	0x62, 0x72, 0x69, 0x63, 0x2e, 0x43, 0x6f, 0x6e, 0x6e, 0x65, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x52,
+	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x1c, 0x82, 0xd3, 0xe4, 0x93, 0x02, 0x16, 0x22,
+	0x0e, 0x2f, 0x76, 0x31, 0x2f, 0x63, 0x6f, 0x6e, 0x6e, 0x65, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x3a,
+	0x04, 0x64, 0x61, 0x74, 0x61, 0x12, 0x61, 0x0a, 0x07, 0x45, 0x78, 0x65, 0x63, 0x75, 0x74, 0x65,
+	0x12, 0x1a, 0x2e, 0x66, 0x61, 0x62, 0x72, 0x69, 0x63, 0x2e, 0x54, 0x72, 0x61, 0x6e, 0x73, 0x61,
+	0x63, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1b, 0x2e, 0x66,
+	0x61, 0x62, 0x72, 0x69, 0x63, 0x2e, 0x54, 0x72, 0x61, 0x6e, 0x73, 0x61, 0x63, 0x74, 0x69, 0x6f,
+	0x6e, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x1d, 0x82, 0xd3, 0xe4, 0x93, 0x02,
+	0x17, 0x22, 0x0f, 0x2f, 0x76, 0x31, 0x2f, 0x74, 0x72, 0x61, 0x6e, 0x73, 0x61, 0x63, 0x74, 0x69,
+	0x6f, 0x6e, 0x3a, 0x04, 0x64, 0x61, 0x74, 0x61, 0x42, 0x0f, 0x5a, 0x0d, 0x66, 0x61, 0x62, 0x72,
+	0x69, 0x63, 0x3b, 0x66, 0x61, 0x62, 0x72, 0x69, 0x63, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f,
+	0x33,
 }
 
-func init() { proto.RegisterFile("fabric.proto", fileDescriptor_284efff686d8e9bf) }
+var (
+	file_fabric_fabric_proto_rawDescOnce sync.Once
+	file_fabric_fabric_proto_rawDescData = file_fabric_fabric_proto_rawDesc
+)
 
-var fileDescriptor_284efff686d8e9bf = []byte{
-	// 584 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x9c, 0x54, 0xcd, 0x6e, 0xd3, 0x40,
-	0x10, 0xc6, 0xcd, 0x8f, 0xe3, 0x69, 0x9a, 0x96, 0x45, 0x02, 0x53, 0x5a, 0xc9, 0xb8, 0x6a, 0x6b,
-	0x71, 0x48, 0xd4, 0x22, 0x81, 0xc4, 0x91, 0x52, 0xa1, 0xa8, 0x6a, 0x11, 0xab, 0x02, 0x82, 0x4b,
-	0x34, 0xb5, 0xa7, 0xae, 0x45, 0xbd, 0x6b, 0xec, 0x4d, 0x4a, 0xae, 0xbc, 0x02, 0xcf, 0xc1, 0x99,
-	0x07, 0xe1, 0x15, 0x78, 0x0c, 0x0e, 0xc8, 0x6b, 0x3b, 0x3f, 0x04, 0x11, 0x89, 0x5b, 0xf6, 0x9b,
-	0x9f, 0x6f, 0xe6, 0xfb, 0xc6, 0x81, 0xf6, 0x25, 0x5e, 0xa4, 0x91, 0xdf, 0x4d, 0x52, 0xa9, 0x24,
-	0x63, 0x81, 0x1c, 0x91, 0xc2, 0xe8, 0xba, 0x5b, 0xc2, 0xa3, 0x83, 0xcd, 0xad, 0x50, 0xca, 0xf0,
-	0x9a, 0x7a, 0x98, 0x44, 0x3d, 0x14, 0x42, 0x2a, 0x54, 0x91, 0x14, 0x59, 0x51, 0xe1, 0x9e, 0x02,
-	0x3b, 0x4f, 0x51, 0x64, 0xe8, 0xe7, 0x28, 0xa7, 0x4f, 0x43, 0xca, 0x14, 0x7b, 0x0a, 0xf5, 0x00,
-	0x15, 0xda, 0x86, 0x63, 0x78, 0xab, 0x87, 0x3b, 0xdd, 0xc5, 0xb6, 0xdd, 0x99, 0xaa, 0x17, 0xa8,
-	0x90, 0xeb, 0x02, 0xf7, 0xdb, 0x0a, 0xac, 0xff, 0x11, 0x61, 0x3b, 0xb0, 0xe6, 0x4b, 0x21, 0x48,
-	0x23, 0x83, 0x28, 0xd0, 0x5d, 0xeb, 0xbc, 0x3d, 0x05, 0xfb, 0x41, 0xce, 0xa8, 0xc6, 0x09, 0xd9,
-	0x2b, 0x8e, 0xe1, 0x75, 0x96, 0x32, 0x9e, 0x8f, 0x13, 0xe2, 0xba, 0x80, 0x3d, 0x84, 0xb6, 0x7f,
-	0x85, 0x91, 0xf0, 0x65, 0x40, 0x79, 0xf3, 0x9a, 0x63, 0x78, 0x16, 0x5f, 0x9d, 0x60, 0xfd, 0x80,
-	0xd9, 0x60, 0xaa, 0x28, 0x26, 0x39, 0x54, 0x76, 0xdd, 0x31, 0xbc, 0x1a, 0xaf, 0x9e, 0x6c, 0x13,
-	0x5a, 0x24, 0x82, 0x44, 0x46, 0x42, 0xd9, 0x0d, 0xa7, 0xe6, 0x59, 0x7c, 0xf2, 0x66, 0x0e, 0xac,
-	0xaa, 0x29, 0xa3, 0xdd, 0x2c, 0xfa, 0xce, 0x40, 0x6c, 0x0b, 0xac, 0x04, 0x53, 0x8c, 0x49, 0x51,
-	0x6a, 0x9b, 0xba, 0x7c, 0x0a, 0xe4, 0x6b, 0xeb, 0xe4, 0x88, 0x84, 0x1a, 0xc4, 0x98, 0xd8, 0x2d,
-	0xdd, 0xa1, 0x3d, 0x01, 0x4f, 0x31, 0x71, 0xdf, 0xc1, 0x9d, 0x39, 0xf9, 0xb3, 0x44, 0x8a, 0x8c,
-	0x18, 0x83, 0x7a, 0x3e, 0xbb, 0x56, 0xaa, 0xc1, 0xf5, 0xef, 0x7c, 0x8b, 0x98, 0xb2, 0x0c, 0xc3,
-	0x42, 0x24, 0x8b, 0x57, 0xcf, 0x3c, 0x5b, 0xbb, 0x55, 0xac, 0x5e, 0x18, 0x71, 0x02, 0xb7, 0x8f,
-	0x26, 0xfa, 0x56, 0xb6, 0x3e, 0x99, 0xb3, 0xd5, 0xfd, 0x9b, 0xc8, 0xd3, 0xa2, 0x19, 0x57, 0xbf,
-	0x1b, 0xd0, 0x99, 0x0f, 0xb0, 0x6d, 0x00, 0xff, 0x0a, 0x85, 0xa0, 0xeb, 0xca, 0x51, 0x8b, 0x5b,
-	0x25, 0xd2, 0x0f, 0xd8, 0x03, 0xb0, 0x86, 0x19, 0xa5, 0x03, 0x81, 0x71, 0x35, 0x6e, 0x2b, 0x07,
-	0xce, 0x30, 0x26, 0x76, 0x1f, 0x5a, 0x32, 0x0d, 0x8b, 0x58, 0x31, 0xb3, 0x29, 0xd3, 0x50, 0x87,
-	0x76, 0xa1, 0x23, 0x48, 0xdd, 0xc8, 0xf4, 0xe3, 0xc0, 0x97, 0xe2, 0x32, 0x0a, 0xb5, 0x63, 0x16,
-	0x5f, 0x2b, 0xd1, 0x23, 0x0d, 0xb2, 0x7d, 0x58, 0x4f, 0x50, 0x29, 0x12, 0x83, 0x18, 0x95, 0x7f,
-	0x45, 0x69, 0x66, 0x37, 0x74, 0x5e, 0xa7, 0x80, 0x4f, 0x4b, 0xd4, 0x0d, 0x81, 0xcd, 0xca, 0xf0,
-	0x5f, 0xf2, 0x2e, 0xdc, 0x6f, 0x6d, 0xf1, 0x7e, 0x1f, 0x79, 0x73, 0x77, 0x9f, 0xdf, 0x27, 0x03,
-	0x68, 0xf6, 0xcf, 0xde, 0xbe, 0x3a, 0x39, 0xde, 0xb8, 0xc5, 0x2c, 0x68, 0xbc, 0x7e, 0x73, 0xcc,
-	0xdf, 0x6f, 0x18, 0x87, 0xbf, 0x0c, 0x30, 0x5f, 0xa2, 0xa2, 0x1b, 0x1c, 0xb3, 0x11, 0x98, 0xe5,
-	0x78, 0x6c, 0xf7, 0xdf, 0x6e, 0x94, 0x16, 0x6e, 0xee, 0x2d, 0x4b, 0x2b, 0x56, 0x74, 0xb7, 0xbe,
-	0xfc, 0xf8, 0xf9, 0x75, 0xe5, 0xae, 0xdb, 0xe9, 0x8d, 0x0e, 0x7a, 0xd3, 0x49, 0x9f, 0x69, 0x43,
-	0xd9, 0x18, 0xcc, 0xe3, 0xcf, 0xe4, 0x0f, 0x15, 0xb1, 0xbd, 0x25, 0x9f, 0x5a, 0x45, 0xbc, 0xbf,
-	0x34, 0xaf, 0x64, 0xde, 0xd6, 0xcc, 0xf7, 0xdc, 0xf5, 0x9c, 0x79, 0xe6, 0x73, 0x29, 0xa8, 0x9f,
-	0xb7, 0x3e, 0x34, 0x8b, 0xfa, 0x8b, 0xa6, 0xfe, 0x07, 0x7a, 0xfc, 0x3b, 0x00, 0x00, 0xff, 0xff,
-	0x09, 0x58, 0x7b, 0x33, 0xc3, 0x04, 0x00, 0x00,
+func file_fabric_fabric_proto_rawDescGZIP() []byte {
+	file_fabric_fabric_proto_rawDescOnce.Do(func() {
+		file_fabric_fabric_proto_rawDescData = protoimpl.X.CompressGZIP(file_fabric_fabric_proto_rawDescData)
+	})
+	return file_fabric_fabric_proto_rawDescData
 }
 
-// Reference imports to suppress errors if they are not otherwise used.
-var _ context.Context
-var _ grpc.ClientConn
-
-// This is a compile-time assertion to ensure that this generated file
-// is compatible with the grpc package it is being compiled against.
-const _ = grpc.SupportPackageIsVersion4
-
-// GatewayClient is the client API for Gateway service.
-//
-// For semantics around ctx use and closing/ending streaming RPCs, please refer to https://godoc.org/google.golang.org/grpc#ClientConn.NewStream.
-type GatewayClient interface {
-	// Connect: create or find a Fabric connection.
-	// Returns the corresponding connection ID.
-	Connect(ctx context.Context, in *ConnectionRequest, opts ...grpc.CallOption) (*ConnectionResponse, error)
-	// Execute: execute a transaction on a Fabric endpoint.
-	// Returns the result of the transaction.
-	Execute(ctx context.Context, in *TransactionRequest, opts ...grpc.CallOption) (*TransactionResponse, error)
+var file_fabric_fabric_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
+var file_fabric_fabric_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
+var file_fabric_fabric_proto_goTypes = []interface{}{
+	(TransactionType)(0),        // 0: fabric.TransactionType
+	(*TransactionRequest)(nil),  // 1: fabric.TransactionRequest
+	(*TransactionData)(nil),     // 2: fabric.TransactionData
+	(*TransactionResponse)(nil), // 3: fabric.TransactionResponse
+	(*ConnectionRequest)(nil),   // 4: fabric.ConnectionRequest
+	(*ConnectionData)(nil),      // 5: fabric.ConnectionData
+	(*ConnectionResponse)(nil),  // 6: fabric.ConnectionResponse
+}
+var file_fabric_fabric_proto_depIdxs = []int32{
+	2, // 0: fabric.TransactionRequest.data:type_name -> fabric.TransactionData
+	0, // 1: fabric.TransactionData.type:type_name -> fabric.TransactionType
+	5, // 2: fabric.ConnectionRequest.data:type_name -> fabric.ConnectionData
+	4, // 3: fabric.Gateway.Connect:input_type -> fabric.ConnectionRequest
+	1, // 4: fabric.Gateway.Execute:input_type -> fabric.TransactionRequest
+	6, // 5: fabric.Gateway.Connect:output_type -> fabric.ConnectionResponse
+	3, // 6: fabric.Gateway.Execute:output_type -> fabric.TransactionResponse
+	5, // [5:7] is the sub-list for method output_type
+	3, // [3:5] is the sub-list for method input_type
+	3, // [3:3] is the sub-list for extension type_name
+	3, // [3:3] is the sub-list for extension extendee
+	0, // [0:3] is the sub-list for field type_name
 }
 
-type gatewayClient struct {
-	cc *grpc.ClientConn
-}
-
-func NewGatewayClient(cc *grpc.ClientConn) GatewayClient {
-	return &gatewayClient{cc}
-}
-
-func (c *gatewayClient) Connect(ctx context.Context, in *ConnectionRequest, opts ...grpc.CallOption) (*ConnectionResponse, error) {
-	out := new(ConnectionResponse)
-	err := c.cc.Invoke(ctx, "/dovetail.fabric.v1.Gateway/Connect", in, out, opts...)
-	if err != nil {
-		return nil, err
+func init() { file_fabric_fabric_proto_init() }
+func file_fabric_fabric_proto_init() {
+	if File_fabric_fabric_proto != nil {
+		return
 	}
-	return out, nil
-}
-
-func (c *gatewayClient) Execute(ctx context.Context, in *TransactionRequest, opts ...grpc.CallOption) (*TransactionResponse, error) {
-	out := new(TransactionResponse)
-	err := c.cc.Invoke(ctx, "/dovetail.fabric.v1.Gateway/Execute", in, out, opts...)
-	if err != nil {
-		return nil, err
+	if !protoimpl.UnsafeEnabled {
+		file_fabric_fabric_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*TransactionRequest); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_fabric_fabric_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*TransactionData); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_fabric_fabric_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*TransactionResponse); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_fabric_fabric_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*ConnectionRequest); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_fabric_fabric_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*ConnectionData); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_fabric_fabric_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*ConnectionResponse); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
 	}
-	return out, nil
-}
-
-// GatewayServer is the server API for Gateway service.
-type GatewayServer interface {
-	// Connect: create or find a Fabric connection.
-	// Returns the corresponding connection ID.
-	Connect(context.Context, *ConnectionRequest) (*ConnectionResponse, error)
-	// Execute: execute a transaction on a Fabric endpoint.
-	// Returns the result of the transaction.
-	Execute(context.Context, *TransactionRequest) (*TransactionResponse, error)
-}
-
-// UnimplementedGatewayServer can be embedded to have forward compatible implementations.
-type UnimplementedGatewayServer struct {
-}
-
-func (*UnimplementedGatewayServer) Connect(ctx context.Context, req *ConnectionRequest) (*ConnectionResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method Connect not implemented")
-}
-func (*UnimplementedGatewayServer) Execute(ctx context.Context, req *TransactionRequest) (*TransactionResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method Execute not implemented")
-}
-
-func RegisterGatewayServer(s *grpc.Server, srv GatewayServer) {
-	s.RegisterService(&_Gateway_serviceDesc, srv)
-}
-
-func _Gateway_Connect_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(ConnectionRequest)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(GatewayServer).Connect(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: "/dovetail.fabric.v1.Gateway/Connect",
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(GatewayServer).Connect(ctx, req.(*ConnectionRequest))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-func _Gateway_Execute_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(TransactionRequest)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(GatewayServer).Execute(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: "/dovetail.fabric.v1.Gateway/Execute",
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(GatewayServer).Execute(ctx, req.(*TransactionRequest))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-var _Gateway_serviceDesc = grpc.ServiceDesc{
-	ServiceName: "dovetail.fabric.v1.Gateway",
-	HandlerType: (*GatewayServer)(nil),
-	Methods: []grpc.MethodDesc{
-		{
-			MethodName: "Connect",
-			Handler:    _Gateway_Connect_Handler,
+	type x struct{}
+	out := protoimpl.TypeBuilder{
+		File: protoimpl.DescBuilder{
+			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
+			RawDescriptor: file_fabric_fabric_proto_rawDesc,
+			NumEnums:      1,
+			NumMessages:   6,
+			NumExtensions: 0,
+			NumServices:   1,
 		},
-		{
-			MethodName: "Execute",
-			Handler:    _Gateway_Execute_Handler,
-		},
-	},
-	Streams:  []grpc.StreamDesc{},
-	Metadata: "fabric.proto",
+		GoTypes:           file_fabric_fabric_proto_goTypes,
+		DependencyIndexes: file_fabric_fabric_proto_depIdxs,
+		EnumInfos:         file_fabric_fabric_proto_enumTypes,
+		MessageInfos:      file_fabric_fabric_proto_msgTypes,
+	}.Build()
+	File_fabric_fabric_proto = out.File
+	file_fabric_fabric_proto_rawDesc = nil
+	file_fabric_fabric_proto_goTypes = nil
+	file_fabric_fabric_proto_depIdxs = nil
 }
